Split auth domain
