+++ src/datasources/auth-api/utils/__tests__/to-signable-siwe-message.spec.ts
@@ -1,5 +1,5 @@
 import { toSignableSiweMessage } from '@/datasources/auth-api/utils/to-signable-siwe-message';
-import { siweMessageBuilder } from '@/domain/auth/entities/__tests__/siwe-message.builder';
+import { siweMessageBuilder } from '@/domain/siwe/entities/__tests__/siwe-message.builder';
 import { faker } from '@faker-js/faker';
 
 describe('toSignableSiweMessage', () => {
