+++ src/domain/siwe/entities/siwe-message.entity.ts
