+++ src/domain/siwe/siwe.repository.interface.ts
@@ -1,0 +1,24 @@
+import { AuthApiModule } from '@/datasources/auth-api/auth-api.module';
+import { SiweRepository } from '@/domain/siwe/siwe.repository';
+import { VerifyAuthMessageDto } from '@/routes/auth/entities/verify-auth-message.dto.entity';
+import { Module } from '@nestjs/common';
+
+export const ISiweRepository = Symbol('ISiweRepository');
+
+export interface ISiweRepository {
+  generateNonce(): Promise<{ nonce: string }>;
+
+  isValidMessage(args: VerifyAuthMessageDto): Promise<boolean>;
+}
+
+@Module({
+  imports: [AuthApiModule],
+  providers: [
+    {
+      provide: ISiweRepository,
+      useClass: SiweRepository,
+    },
+  ],
+  exports: [ISiweRepository],
+})
+export class SiweRepositoryModule {}
